made a gcode module
